requirements.txt
kubernetes
google-cloud-storage
python-dotenv
google-cloud-pubsub
typing